Breakup reqres module
